requirements.txt
asgiref==3.5.0
backports.zoneinfo==0.2.1
Django==4.0.2
django-location-field==2.1.0
django-map-widgets==0.3.3
gunicorn==20.1.0
six==1.16.0
sqlparse==0.4.2
whitenoise==5.3.0
